IconButton: barrel structure refactor
