+++ client/common/IconButton/index.ts
@@ -1,0 +1,3 @@
+export { default } from './IconButton';
+export type { IconButtonProps } from './IconButton.types';
+export { ButtonWrapper } from './IconButton.styles';
